requirements.txt
streamlit==1.45.1
pandas==2.2.3
Pillow==10.4.0
watchdog==6.0.0 